+++ src/main/java/org/launchcode/books2borrow/data/MessageRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import java.util.List;
 
 public interface MessageRepository extends JpaRepository<Message, Long> {
-    List<Message> findBySenderIdAndRecipientId(Long senderId, Long recipientId);
+    List<Message> findBySenderIdAndRecipientId(int senderId, int recipientId);
 }
